Style change, Updated Cart, and more responsive features
Reverted to previous style, added more features to car, and made pages more responsive and look better on mobile.

+++ salonsite/salon/views.py
@@ -122,9 +122,10 @@ def change_password(request):
     })
 
 def view_cart(request):
+    all_products = Product.objects.all()
     cart_items = CartItem.objects.filter(user=request.user)
     total_price = sum(item.product.cost * item.quantity for item in cart_items)
-    return render(request, 'cart.html', {'cart_items': cart_items, 'total_price': total_price})
+    return render(request, 'cart.html', {'cart_items': cart_items, 'total_price': total_price,'products':all_products })
 
 def add_to_cart(request, product_id):
     product = Product.objects.get(id=product_id)
@@ -133,9 +134,14 @@ def add_to_cart(request, product_id):
     cart_item.save()
     return redirect('view_cart')
 
+
 def remove_from_cart(request, item_id):
     cart_item = CartItem.objects.get(id=item_id)
-    cart_item.delete()
+    if cart_item.quantity <= 1:
+        cart_item.delete()
+    else:
+        cart_item.quantity= cart_item.quantity -1
+        cart_item.save()
     return redirect('view_cart')
 
 def place_order(request):
